Added WASI memory64 support.

+++ wasm/src/org.graalvm.wasm/src/org/graalvm/wasm/predefined/BuiltinModule.java
@@ -155,7 +155,7 @@ public abstract class BuiltinModule {
         instance.symbolTable().exportFunction(function.index(), exportName);
     }
 
-    protected void importMemory(WasmInstance instance, String importModuleName, String memoryName, int initSize, int maxSize, boolean is64Bit) {
+    protected void importMemory(WasmInstance instance, String importModuleName, String memoryName, int initSize, long maxSize, boolean is64Bit) {
         instance.symbolTable().importMemory(importModuleName, memoryName, initSize, maxSize, is64Bit);
     }
